refactor(ai-teacher): use server components for data fetching
