+++ src/app/(app)/ai-teacher/components/chat-area.tsx
@@ -4,6 +4,7 @@ import { useState, useEffect, useRef, useCallback } from 'react';
 import { AiTeacher, AiConversation, AiMessage } from '@/lib/types/ai';
 import { ChatMessage } from './chat-message';
 import { MessageInput } from './message-input';
+import { sendMessage as sendMessageAction } from '../chat/actions';
 
 interface ChatAreaProps {
   teacher: AiTeacher;
@@ -56,13 +57,12 @@ export function ChatArea({
     messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
   };
 
-  const sendMessage = async (content: string) => {
+  const handleMessageSend = async (content: string) => {
     if (!content.trim()) return;
 
     setSendingMessage(true);
 
     // Optimistic update: add user message immediately to UI
-    // TODO: add sending indicator
     const userMessage: AiMessage = {
       id: `temp-${Date.now()}`,
       role: 'user',
@@ -73,50 +73,25 @@ export function ChatArea({
     setMessages(prev => [...prev, userMessage]);
 
     try {
-      const response = await fetch('/api/ai/chat', {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({
-          teacherId: teacher.id,
-          message: content,
-          conversationId: conversation?.id,
-        }),
-      });
+      const result = await sendMessageAction(teacher.id, content, conversation?.id);
 
-      const data = await response.json();
-
-      if (response.ok) {
-        // If this was a new conversation, notify parent
-        if (!conversation && data.conversationId) {
-          const newConversation: AiConversation = {
-            id: data.conversationId,
-            title: content.substring(0, 50) + (content.length > 50 ? '...' : ''),
-            created_at: new Date().toISOString(),
-            updated_at: new Date().toISOString(),
-          };
-          onConversationCreated(newConversation);
-        }
-
-        // Add AI response
-        const aiMessage: AiMessage = {
-          id: `ai-${Date.now()}`,
-          role: 'assistant',
-          content: data.message,
-          created_at: new Date().toISOString(),
-        };
-
-        setMessages(prev => [...prev.slice(0, -1), {
-          ...userMessage,
-          id: `user-${Date.now()}`,
-        }, aiMessage]);
-      } else {
-        // Remove the temporary user message on error
-        setMessages(prev => prev.slice(0, -1));
-        // TODO: handle error gracefully in UI
-        alert(data.error || 'Failed to send message');
+      // If this was a new conversation, notify parent
+      if (!conversation && result.newConversation) {
+        onConversationCreated(result.newConversation);
       }
+
+      // Add AI response
+      const aiMessage: AiMessage = {
+        id: `ai-${Date.now()}`,
+        role: 'assistant',
+        content: result.message,
+        created_at: new Date().toISOString(),
+      };
+
+      setMessages(prev => [...prev.slice(0, -1), {
+        ...userMessage,
+        id: `user-${Date.now()}`,
+      }, aiMessage]);
     } catch (error) {
       // Remove the temporary user message on error
       setMessages(prev => prev.slice(0, -1));
@@ -212,7 +187,7 @@ export function ChatArea({
 
       {/* Message Input */}
       <MessageInput
-        onSend={sendMessage}
+        onSend={handleMessageSend}
         disabled={sendingMessage}
         placeholder={`Ask ${teacher.name} anything about ${teacher.subject}...`}
       />
